# Use rust backend to speed up decoding / encoding