--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -58,13 +58,6 @@ jobs:
       - image: circleci/python:3.7
         environment:
           TOXENV: py37-core
-  pypy3-core:
-    <<: *common
-    docker:
-      - image: pypy:3-5.8.0
-        environment:
-          TOXENV: pypy3-core
-
   py35-doctest:
     <<: *common
     docker:
@@ -83,12 +76,6 @@ jobs:
       - image: circleci/python:3.7
         environment:
           TOXENV: py37-doctest
-  pypy3-doctest:
-    <<: *common
-    docker:
-      - image: pypy:3-5.8.0
-        environment:
-          TOXENV: pypy3-doctest
 
   lint:
     <<: *common
@@ -104,11 +91,9 @@ workflows:
       - py35-core
       - py36-core
       - py37-core
-      - pypy3-core
 
       - py35-doctest
       - py36-doctest
       - py37-doctest
-      - pypy3-doctest
 
       - lint