--- a/setup.py
+++ b/setup.py
@@ -8,9 +8,9 @@ from setuptools import (
 
 extras_require = {
     'test': [
-        "pytest==3.3.2",
+        "pytest==5.4.3",
         "tox>=2.9.1,<3",
-        "hypothesis==3.56.5",
+        "hypothesis==5.19.0",
     ],
     'lint': [
         "flake8==3.4.1",