--- a/tox.ini
+++ b/tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist =
-    py{35,36,37,py3}-core
-    py{35,36,37,py3}-doctest
+    py{35,36,37}-core
+    py{35,36,37}-doctest
     lint
 
 [flake8]
@@ -18,7 +18,6 @@ basepython =
     py35: python3.5
     py36: python3.6
     py37: python3.7
-    pypy3: pypy3
 extras=
     test
 whitelist_externals=make