### What was wrong

#### Current Performance

```
$ python tests/speed.py 
Block serializations / sec: 3224.00
Block deserializations / sec: 1907.96
TX serializations / sec: 41606.69
TX deserializations / sec: 23015.79
```

### How was it fixed

#### Performance with this PR

```
$ python tests/speed.py 
Block serializations / sec: 6188.00
Block deserializations / sec: 2194.96
TX serializations / sec: 72538.08
TX deserializations / sec: 28787.28
```

These numbers :point_up: where more impressive during development (and they still are depending on the specific benchmark) but I think I've lost some performance due to handling various edge cases. It's still faster though. That said, I believe there is still lots of room in [`rusty-rlp`](https://github.com/cburgdorf/rusty-rlp) to make things faster.

### TODO:

- [x] decide what to do about pypy
- [ ] decide what to do about ignored edge case
- [x] Ensure `rusty-rlp` produces mac and windows builds
- [x] decide if we can live without the recursive cache (it's disabled by default anyway)
- [x] Ensure [`rusty-rlp`](https://github.com/cburgdorf/rusty-rlp) has CI properly setup