@cburgdorf 2020-08-17
I dropped pypy support in this PR. I made sure that `rusty-rlp` [is released](https://pypi.org/project/rusty-rlp/0.1.10/#files) with broad platform support, precisely: Linux, Mac, Windows with Python versions 3.5, 3.6, 3.7 and 3.8.

My current thinking is that it would be relatively harmless to cut an **alpha** version of `v2` and start using that in Py-EVM and Trinity (~~assuming we don't find any big blockers~~, assuming we fixed all [outstanding blockers](https://github.com/ethereum/py-evm/pull/1951)). That would still give us time to make follow up cleanups before we cut an actual stable `v2` release. I realize that this is one more alpha dependency for Py-EVM / Trinity (which we want to get rid of for Trinity's beta) but there has to be some cost to unlocking a performance improvement quickly :)

Would that sound reasonable to you? @pipermerriam @carver 

Speaking of the performance implications of this change, it occured to me that this has the potential to speed things up on various levels: Block processing, Db writes / reads, networking etc. So it may even contribute to our recent vendetta of cutting down on things that hold the event loop hostage for too long.