@pipermerriam 2020-08-18
It's worth spending a minute figuring out what we want to formally document as public API and what we want to relegate to private APIs before the `v2` since it is a chance to make breaking changes and who doesn't love breaking changes :smiling_imp: 

I don't have anything specific in mind so there might be nothing here...

You did mention that the recursive cache was a problem...