@cburgdorf 2020-08-18
Another quick update here:

1. I have opened a [PR against Py-EVM](https://github.com/ethereum/py-evm/pull/1951) to make sure there aren't any issues that went unnoticed. It works! And the nice thing is that one can even observe a performance boost in the benchmarks. (I ran this a few times locally just toggling between a virtualenv with the new rlp and one with the old rlp and the results are consistent)

Current:

```
Starting benchmark: Simple value transfer

Sending to existing address

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     5.738      |      142       |     24.749     |       1        |       0.174        |   2,982,000    |  519,731.556   |
|     homestead     |     6.308      |      142       |     22.512     |       1        |       0.159        |   2,982,000    |  472,756.675   |
| tangerine-whistle |     6.299      |      142       |     22.543     |       1        |       0.159        |   2,982,000    |  473,398.113   |
|  spurious-dragon  |     6.290      |      142       |     22.577     |       1        |       0.159        |   2,982,000    |  474,118.907   |
|     byzantium     |     6.300      |      142       |     22.540     |       1        |       0.159        |   2,982,000    |  473,336.865   |
|    petersburg     |     6.282      |      142       |     22.604     |       1        |       0.159        |   2,982,000    |  474,682.378   |
|     istanbul      |     6.362      |      142       |     22.319     |       1        |       0.157        |   2,982,000    |  468,700.784   |
|   muir-glacier    |     6.299      |      142       |     22.544     |       1        |       0.159        |   2,982,000    |  473,430.261   |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     49.877     |      1136      |       -        |       8        |         -          |   23,856,000   |       -        |
|        Avg        |     6.235      |      142       |     22.776     |       1        |       0.160        |   2,982,000    |  478,296.726   |
================================================================================================================================================

```

New:

```
Starting benchmark: Simple value transfer

Sending to existing address

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     4.398      |      142       |     32.290     |       1        |       0.227        |   2,982,000    |  678,089.862   |
|     homestead     |     4.400      |      142       |     32.273     |       1        |       0.227        |   2,982,000    |  677,736.711   |
| tangerine-whistle |     4.400      |      142       |     32.276     |       1        |       0.227        |   2,982,000    |  677,793.365   |
|  spurious-dragon  |     4.400      |      142       |     32.273     |       1        |       0.227        |   2,982,000    |  677,732.111   |
|     byzantium     |     4.441      |      142       |     31.971     |       1        |       0.225        |   2,982,000    |  671,399.438   |
|    petersburg     |     4.864      |      142       |     29.191     |       1        |       0.206        |   2,982,000    |  613,017.025   |
|     istanbul      |     4.861      |      142       |     29.215     |       1        |       0.206        |   2,982,000    |  613,515.807   |
|   muir-glacier    |     4.846      |      142       |     29.302     |       1        |       0.206        |   2,982,000    |  615,338.591   |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     36.610     |      1136      |       -        |       8        |         -          |   23,856,000   |       -        |
|        Avg        |     4.576      |      142       |     31.030     |       1        |       0.219        |   2,982,000    |  651,630.883   |
================================================================================================================================================

```

Notice the higher throughput (tx / second and gas / second) that shows an increase of `1.36x`.

2. I have also opened a PR against Trinity and unfortunately there is [one case with *non-strict* encoding](https://app.circleci.com/pipelines/github/ethereum/trinity/7246/workflows/73f3758f-6edf-4c1d-a48a-8e192ac9d0e0/jobs/276808) that I have yet to investigate.

3. I also noticed that `p2p` does in fact [use the `recursive_cache` feature](https://github.com/ethereum/trinity/pull/1988/files#diff-bbb0e89938e49714d945f9d70657b870L64) that is currently unavailable. I *think* that I can make a change to `rusty-rlp` so that we can continue to have that functionality.