@cburgdorf 2020-08-18
@pipermerriam Sorry, I wrote my post before I noticed you had made a comment.

>It's worth spending a minute figuring out what we want to formally document as public API and what we want to relegate to private APIs before the v2 since it is a chance to make breaking changes and who doesn't love breaking changes

Absolutely. By I think by cutting a `v2` **alpha** we can defer the decision just a bit and still get the performance wins immediately.

>You did mention that the recursive cache was a problem...

Yep, my comment above goes into that. I still want to investigate if I can make sure we keep that functionality.